requirements.txt
astroid==1.5.3
colorama==0.3.9
coverage==4.4.1
dj-database-url==0.4.2
Django==2.1.3
django-environ==0.4.4
gunicorn==19.7.1
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==1.7.2
pytz==2017.2
requests==2.18.4
selenium==3.5.0
six==1.10.0
whitenoise==3.3.0
wrapt==1.10.11
django-crispy-forms==1.7.2
psycopg2==2.7.7
docker-compose
djangorestframework
markdown
django-filter
django-debug-toolbar